create-slurm-security-groups/requirements.txt
aws-cdk-lib==2.111.0
boto3
colored
constructs>=10.0.0,<11.0.0
